+++ app/src/main/java/jp/developer/retia/frozenword/MainActivity.kt
@@ -21,7 +21,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import dagger.hilt.android.AndroidEntryPoint
-import jp.developer.retia.frozenword.ui.activity.HabbitsActivity
+import jp.developer.retia.frozenword.ui.habbits.HabbitsActivity
 import kotlinx.coroutines.InternalCoroutinesApi
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
